services/flight-cache/requirements.txt
flask
flask_cors
Flask==2.3.3
psycopg2-binary==2.9.9
